# api/swagger.yml
---
swagger: '2.0'
info:
  version: 1.0.0
  title: Swagger Demo
  description: For API design, documentation and testing
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
host: localhost:3000
basePath: "/api/v1"
tags:
- name: Users
  description: API for users in the system
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
  /users:
    post:
      tags:
      - Users
      summary: "Add a new user"
      description: "Create new user in system"
      consumes:
      - "application/json"
      produces:
      - application/json
      parameters:
      - in: "body"
        name: "body"
        schema:
          "$ref": "#/definitions/User"
      responses:
        '200':
          description: New user is created
          schema:
            "$ref": "#/definitions/User"
    get:
      tags:
      - Users
      summary: Get all users in system
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/Users"
  "/users/{userId}":
    parameters:
    - name: userId
      in: path
      required: true
      description: ID of user that we want to find
      type: string
    get:
      tags:
      - Users
      summary: Get user with given ID
      responses:
        '200':
          description: User is found
          schema:
            "$ref": "#/definitions/User"
    delete:
      summary: Delete user with given ID
      tags:
      - Users
      responses:
        '200':
          description: User is deleted
          schema:
            "$ref": "#/definitions/User"
    put:
      summary: Update user with give ID
      tags:
      - Users
      parameters:
      - name: user
        in: body
        description: User with new values of properties
        schema:
          "$ref": "#/definitions/User"
      responses:
        '200':
          description: User is updated
          schema:
            "$ref": "#/definitions/User"
definitions:
  User:
    required:
    - email
    - _id
    properties:
      _id:
        type: string
        uniqueItems: true
        example: '5aad0b6c5982721361589deb'
      email:
        type: string
        uniqueItems: true
      lastName:
        type: string
      firstName:
        type: string
  Users:
    type: array
    items:
      "$ref": "#/definitions/User"